- 修复 GroupBy 类型转换错误；#186 - 修复 .ToList(a => new DTO(a.id)) 报 未将对象引用设置到对象的实例 问题； #187 - 修复 update语句，二元运算解析出错； #184

+++ FreeSql.Tests/FreeSql.Tests/UnitTest1.cs
@@ -698,7 +698,7 @@ namespace FreeSql.Tests
                  .ToSql(a => new NewsArticleDto
                  {
                      ArticleTitle = a.Key,
-                      ChannelId = a.Sum(a.Value.Item1.OptionsEntity04)
+                      ChannelId = (int)a.Sum(a.Value.Item1.OptionsEntity04)
                  });
 
             var testgrpsql2 = g.sqlite.Select<TaskBuild>()
